requirements.txt
Adafruit_DHT
schedule
adafruit-circuitpython-neopixel
random
paho-mqtt
